Change imports and move dashboards to mango_time_series
