+++ mango_time_series/mango_time_series/dashboards/time_series_app.py
@@ -1,37 +1,38 @@
 import streamlit as st
-from mango_base.mango.dashboards.time_series_utils.constants import (
+from statsforecast import StatsForecast
+
+from .time_series_utils.constants import (
     SELECT_AGR_TMP_DICT,
     model_context,
 )
-from mango_base.mango.dashboards.time_series_utils.constants import default_models
-from mango_base.mango.dashboards.time_series_utils.data_loader import load_data
-from mango_base.mango.dashboards.time_series_utils.data_processing import (
+from .time_series_utils.constants import default_models
+from .time_series_utils.data_loader import load_data
+from .time_series_utils.data_processing import (
     process_data,
     convert_df,
     aggregate_to_input_cache,
     render_script,
 )
-from mango_base.mango.dashboards.time_series_utils.file_uploader import (
+from .time_series_utils.file_uploader import (
     upload_files,
     manage_files,
 )
-from mango_base.mango.dashboards.time_series_utils.ui_components import (
+from .time_series_utils.ui_components import (
     plot_time_series,
     setup_sidebar,
     plot_forecast,
     plot_error_visualization,
     adapt_values_based_on_series_length,
 )
-from mango_base.mango.dashboards.time_series_utils.ui_text_catala import (
+from .time_series_utils.ui_text_catala import (
     UI_TEXT as UI_TEXT_CATALA,
 )
-from mango_base.mango.dashboards.time_series_utils.ui_text_en import (
+from .time_series_utils.ui_text_en import (
     UI_TEXT as UI_TEXT_EN,
 )
-from mango_base.mango.dashboards.time_series_utils.ui_text_es import (
+from .time_series_utils.ui_text_es import (
     UI_TEXT as UI_TEXT_ES,
 )
-from statsforecast import StatsForecast
 
 
 # from statsmodels.tsa.seasonal import STL
