+++ mango_time_series/mango_time_series/dashboards/time_series_utils/data_processing.py
@@ -3,8 +3,9 @@ import importlib.resources as pkg_resources
 import jinja2
 import pandas as pd
 import streamlit as st
-from mango_base.mango.dashboards.time_series_utils.constants import all_imports
-from mango_time_series.mango.utils.processing import aggregate_to_input
+from mango_time_series.utils.processing import aggregate_to_input
+
+from .constants import all_imports
 
 
 @st.cache_data
