+++ mango_time_series/mango_time_series/dashboards/time_series_utils/forecast_template.py.j2
