+++ mango_time_series/mango_time_series/dashboards/time_series_utils/ui_text_catala.py
