+++ mango_time_series/mango_time_series/tests/example_timeseries_creation_polars.py
@@ -2,7 +2,7 @@ from datetime import date
 
 import numpy as np
 import polars as pl
-from mango_base.mango.logging import log_time
+from mango.logging import log_time
 
 np.random.seed(42)
